backend/requirements.txt
# Gaplytics Backend Dependencies
# AI: AWS Bedrock (Claude 3 Haiku) via Bearer token — no boto3 needed
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pymupdf>=1.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mangum>=0.17.0
